PC3_Interbank/requirements.txt
asgiref==3.8.1
channels==4.2.2
chardet==5.2.0
Django==5.2.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
reportlab==4.4.1
sqlparse==0.5.3
tzdata==2025.2
google-adk==1.3.0
litellm==1.72.4